% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_f24.R
\name{collect_f1_match_id}
\alias{collect_f1_match_id}
\title{Collect all matches id from a team specified in parameter.
The collected \code{match_id} can then be used in \code{collect_f24}.}
\usage{
collect_f1_match_id(
  user,
  passwd,
  season,
  team_id,
  competition = 24,
  root_url = "http://omo.akamai.opta.net/competition.php"
)
}
\arguments{
\item{user}{User to connect to the \code{root_url} account\verb{;}}

\item{passwd}{Password of the \code{user} to use.}

\item{season}{Season of the games to collect. If set, nor \code{match_id} nor \code{matches_id} are read.}

\item{team_id}{Team to collect the matches from.}

\item{root_url}{The url to collect the f24 file. By default it is set to \url{http://omo.akamai.opta.net}.}
}
\value{
A tibble made from four columns: a \code{match_id}, \code{home_team}, \code{away_team},
\dontrun{
matches <- collect_f1_match_id(user="xxxxx", passwd = "xxxxx", team_id="xxxx", season=2020)

# collect 1st match
collect_f24(user="xxxxx", passwd = "xxxxx", season=2020, matches_id = matches$match_id[1], output_dir = "my_path/to/folder")

collect all matches
collect_f24(user="xxxxx", passwd = "xxxxx", season=2020, matches_id = matches$match_id, output_dir = "my_path/to/folder")

}
}
\description{
Collect all matches id from a team specified in parameter.
The collected \code{match_id} can then be used in \code{collect_f24}.
}
