% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_f24.R
\name{collect_f1_teams_id}
\alias{collect_f1_teams_id}
\title{Collect teams' id for a \code{season} and a \code{competition}.}
\usage{
collect_f1_teams_id(
  user,
  passwd,
  competition = 24,
  season,
  root_url = "http://omo.akamai.opta.net/competition.php"
)
}
\arguments{
\item{user}{User to connect to the \code{root_url} account\verb{;}}

\item{passwd}{Password of the \code{user} to use.}

\item{competition}{Competition id. Default set to \code{24}.}

\item{season}{Season of the games to collect. If set, nor \code{match_id} nor \code{matches_id} are  read.}

\item{root_url}{The url to collect the f24 file. By default it is set to \url{http://omo.akamai.opta.net/competition.php}.}
}
\value{
A tibble containing two columns: a \code{team_id} and a \code{label}
}
\description{
Collect teams' id for a \code{season} and a \code{competition}.
}
\examples{
\dontrun{
iniestat::collect_f1_teams_id(user = "MYUSER", passwd = "MYPASSWD", season = 2020)
}
}
