% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_f24.R
\name{collect_f24}
\alias{collect_f24}
\title{Collect all opta available files for a specified season.}
\usage{
collect_f24(
  user,
  passwd,
  output_dir,
  season = NULL,
  match_id = NULL,
  matches_id = NULL,
  root_url = "http://omo.akamai.opta.net",
  store_f7 = TRUE,
  ...
)
}
\arguments{
\item{user}{User to connect to the \code{root_url} account\verb{;}}

\item{passwd}{Password of the \code{user} to use.}

\item{output_dir}{Output directory to write files to.}

\item{season}{Season of the games to collect. If set, nor \code{match_id} nor \code{matches_id} are read.}

\item{match_id}{Specific game id. Do not use with \code{season} set.}

\item{matches_id}{Specific games ids. Do not use with \code{season} set. If set, will overwrite \code{match_id}.}

\item{root_url}{The URL to collect the f24 file. By default it is set to \url{http://omo.akamai.opta.net}.}

\item{store_f7}{Store as well corresponding F7 file. Default set to \code{TRUE}.}

\item{...}{Additional arguments for the underlying functions.}
}
\value{
Save to severals .xml files the collected .xml files.
}
\description{
Collect all opta available files for a specified season.
}
\examples{
 
\dontrun{
# if you want all the available game of the season `season`
collect_f24(user="XX12345", passwd="mypasswd", season=2020)

# if you want the game of id `match_id`
collect_f24(user="XX12345", passwd="mypasswd", match_id="7890")

# if you want several games, use `matches_id` as
collect_f24(user="XX12345", passwd="mypasswd", match_id=c("7890", "7891", "7177"))

# if you need to specify the root url you can set
collect_f24(user="XX12345", passwd="mypasswd", season=2020, root_url="https://my_new_root_url")
}
}
