% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f24.R
\name{read_f24}
\alias{read_f24}
\title{Read OPTA-F24 from a single file OR all files from a directory.}
\usage{
read_f24(file_path = NULL, dir_path = NULL)
}
\arguments{
\item{file_path}{the OPTA-f24 file to parse. The extension has to be \code{.xml}.}

\item{dir_path}{the OPTA-f24 file to parse. The files in the dir have to be f24 with an \code{.xml} extension.}
}
\value{
A tibble wich contains the events of the file specified in \code{file_path} or ones from \code{dir_path} all together.
}
\description{
Read OPTA-F24 from a single file OR all files from a directory.
}
\examples{
 
# use as follows
# my_events_tibble <- read_f24('my_f24_file.xml')
}
\seealso{
read_opt
}
