% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_f24.R
\name{read_f24_single_file_}
\alias{read_f24_single_file_}
\title{Read OPTA-F24 from a single file.
Internal use only.}
\usage{
read_f24_single_file_(file_path)
}
\value{
A tibble wich contains the evnets of the file specified in \code{file_path}.
}
\description{
Read OPTA-F24 from a single file.
Internal use only.
}
